updated hash_block()
added sort_keys argument to sort the keys of the dictionary before dumping it to a string. This allows the same dictionary to always lead to the same string.

OrderedDict was imported from collections library so as to ensure transactions have the same order as well.

OrderedDict takes a list of tuples comprised of key:value pairs.